+++ Runtime/Components/Cinemachine3rdPersonFollow.cs
@@ -88,7 +88,8 @@ namespace Cinemachine
             var dampedTargetPos = Quaternion.Inverse(curState.RawOrientation) 
                 * (targetPos - prevTargetPos);
             if (deltaTime >= 0)
-                dampedTargetPos = Damper.Damp(dampedTargetPos, Damping, deltaTime);
+                dampedTargetPos = VirtualCamera.DetachedFollowTargetDamp(
+                    dampedTargetPos, Damping, deltaTime);
             dampedTargetPos = prevTargetPos + curState.RawOrientation * dampedTargetPos;
 
             // Get target rotation (worldspace)