+++ Runtime/Components/CinemachineComposer.cs
@@ -238,9 +238,12 @@ namespace Cinemachine
 
                 // First force the previous rotation into the hard bounds, no damping,
                 // then  move it through the soft zone, with damping
-                RotateToScreenBounds(
-                    ref curState, mCache.mFovHardGuideRect, TrackedPoint,
-                    ref rigOrientation, mCache.mFov, mCache.mFovH, -1);
+                if (deltaTime < 0 || VirtualCamera.LookAtTargetAttachment > 1 - Epsilon)
+                {
+                    RotateToScreenBounds(
+                        ref curState, mCache.mFovHardGuideRect, TrackedPoint,
+                        ref rigOrientation, mCache.mFov, mCache.mFovH, -1);
+                }
                 RotateToScreenBounds(
                     ref curState, mCache.mFovSoftGuideRect, TrackedPoint,
                     ref rigOrientation, mCache.mFov, mCache.mFovH, deltaTime);
@@ -432,8 +435,10 @@ namespace Cinemachine
             // Apply damping
             if (deltaTime >= 0 && VirtualCamera.PreviousStateIsValid)
             {
-                rotToRect.x = Damper.Damp(rotToRect.x, m_VerticalDamping, deltaTime);
-                rotToRect.y = Damper.Damp(rotToRect.y, m_HorizontalDamping, deltaTime);
+                rotToRect.x = VirtualCamera.DetachedLookAtTargetDamp(
+                    rotToRect.x, m_VerticalDamping, deltaTime);
+                rotToRect.y = VirtualCamera.DetachedLookAtTargetDamp(
+                    rotToRect.y, m_HorizontalDamping, deltaTime);
             }
 
             // Rotate