+++ Runtime/Experimental/CinemachineNewVirtualCamera.cs
@@ -144,6 +144,9 @@ namespace Cinemachine
         /// <param name="deltaTime">Delta time for time-based effects (ignore if less than 0)</param>
         override public void InternalUpdateCameraState(Vector3 worldUp, float deltaTime)
         {
+            FollowTargetAttachment = 1;
+            LookAtTargetAttachment = 1;
+
             // Initialize the camera state, in case the game object got moved in the editor
             m_State = PullStateFromVirtualCamera(worldUp, ref m_Lens);
 
@@ -194,6 +197,9 @@ namespace Cinemachine
             ref CameraState state, Vector3 worldUp, float deltaTime)
         {
             UpdateComponentCache();
+
+            // Extensions first
+            InvokePrePipelineMutateCameraStateCallback(this, ref state, deltaTime);
 
             // Apply the component pipeline
             for (CinemachineCore.Stage stage = CinemachineCore.Stage.Body;